xsv-taichi-001: repair G1 entrypoint and runtime boundary
