+++ PACKAGE_QA_STATUS.json
@@ -34,8 +34,8 @@
     "current_authority_consistency_test_count": 6,
     "val001_synthetic_test_count": 58,
     "static_gate_count": 38,
-    "source_manifest_file_count": 282,
-    "source_manifest_aggregate_sha256": "5a51b54793ed498562664dd2ed87a0d5527b344fdc9d81265ee5cf888f64ef01",
+    "source_manifest_file_count": 283,
+    "source_manifest_aggregate_sha256": "7a0831e8a633305ecb3cb7ecd859a0c3f0a2de89b8544cd6e4c142d24fa910f5",
     "historical_baseline_integrity": "PASS",
     "active_governing_change_contract": "PASS",
     "release_integrity": "PASS",
