+++ SOURCE_PACKAGE_MANIFEST.json
@@ -26,10 +26,10 @@
     "Make/linux*/ build products",
     "deterministically generated case dictionaries and case/0 fields"
   ],
-  "file_count": 282,
-  "total_uncompressed_bytes_excluding_manifest": 3236492,
-  "aggregate_source_sha256": "5a51b54793ed498562664dd2ed87a0d5527b344fdc9d81265ee5cf888f64ef01",
-  "aggregate_sha256": "5a51b54793ed498562664dd2ed87a0d5527b344fdc9d81265ee5cf888f64ef01",
+  "file_count": 283,
+  "total_uncompressed_bytes_excluding_manifest": 3241006,
+  "aggregate_source_sha256": "7a0831e8a633305ecb3cb7ecd859a0c3f0a2de89b8544cd6e4c142d24fa910f5",
+  "aggregate_sha256": "7a0831e8a633305ecb3cb7ecd859a0c3f0a2de89b8544cd6e4c142d24fa910f5",
   "strategy_source_path": "docs/source_strategy/espresso_puck_modeling_and_simulation_strategy_v1_4.md",
   "strategy_source_sha256": "3e3b96b39f81c5e2df5c2dbe42103a4a5a58f928b9e5409bbcaeb861761377b9",
   "files": {
@@ -1184,9 +1184,9 @@
       "git_mode": "100755"
     },
     "scripts/xsv_taichi_001.py": {
-      "sha256": "9250d6db0d113b4c62400e1e19804762d3b9820e22005be5711565400ca93997",
-      "bytes": 7309,
-      "git_mode": "100644"
+      "sha256": "b5e12f40cdfc6245b8c1192abf66b4db4d269629ee2f81eb3800c0f012cdf7b0",
+      "bytes": 486,
+      "git_mode": "100755"
     },
     "solver/espressoWholePullFoam/Make/files": {
       "sha256": "70aa84e2e2cbd37a9c25217fee498c234816cf41fb57b5f24c41a0e8495f7d83",
@@ -1414,8 +1414,8 @@
       "git_mode": "100644"
     },
     "tests/test_xsv_taichi_001.py": {
-      "sha256": "dfa3312c5d49c24bf0f2fe194e35b6e7ecac7066bb4862b628fe67d7ead972b3",
-      "bytes": 4740,
+      "sha256": "f7e2cf71145092e6f7bf1a7f0b1d42b6872325ead118f842157e0d484b99830b",
+      "bytes": 7013,
       "git_mode": "100644"
     },
     "verification/cases/xsv_taichi_001/XSV_TAICHI_001_CASE_MATRIX.csv": {
@@ -1434,13 +1434,18 @@
       "git_mode": "100644"
     },
     "verification/cases/xsv_taichi_001/XSV_TAICHI_001_STAGE_AUTHORIZATION.json": {
-      "sha256": "17dbbb8567a2d64f6cb2b95f2e2ba63a2c5e419c1b3134b856fa2c4eaa175f0e",
-      "bytes": 5489,
+      "sha256": "cc9fb81f7730af94538307992e9226692c103f6a706a60742d2e7fb063c73a35",
+      "bytes": 6912,
       "git_mode": "100644"
     },
     "verification/cases/xsv_taichi_001/lbm/README.md": {
       "sha256": "781d543e99729158559aabfb35779750a83248a2cd8ae253a37afa838d14dcb4",
       "bytes": 311,
+      "git_mode": "100644"
+    },
+    "verification/cases/xsv_taichi_001/xsv_taichi_001_runtime.py": {
+      "sha256": "7dd9f331da30bdbb77469cf62516868de5f3999b6280c19e1ff2da9be33778bd",
+      "bytes": 7641,
       "git_mode": "100644"
     }
   }
