+++ tests/test_xsv_taichi_001.py
@@ -1,6 +1,10 @@
 import csv
+import hashlib
 import json
+import os
 from pathlib import Path
+import re
+import stat
 import unittest
 
 
@@ -45,6 +49,45 @@ class XSVTaichi001Tests(unittest.TestCase):
         )
         self.assertEqual(
             protocol["claim_ceiling"]["physical_validation"], "NOT_ESTABLISHED"
+        )
+
+        launcher_path = ROOT / "scripts/xsv_taichi_001.py"
+        runtime_path = CASE_ROOT / "xsv_taichi_001_runtime.py"
+        self.assertTrue(launcher_path.is_file())
+        self.assertTrue(runtime_path.is_file())
+        self.assertTrue(os.access(launcher_path, os.X_OK))
+        self.assertTrue(launcher_path.stat().st_mode & stat.S_IXUSR)
+        launcher = launcher_path.read_text(encoding="utf-8")
+        runtime = runtime_path.read_text(encoding="utf-8")
+        self.assertEqual(launcher.splitlines()[0], "#!/usr/bin/env python3")
+        self.assertIn("runpy.run_path", launcher)
+        self.assertIn('"verification"', launcher)
+        self.assertIn('"cases"', launcher)
+        self.assertIn('"xsv_taichi_001"', launcher)
+        self.assertIn('"xsv_taichi_001_runtime.py"', launcher)
+        forbidden_import = re.compile(
+            r"^\s*(?:from|import)\s+(?:numpy|pandas|scipy|taichi)(?:\s|\.|$)",
+            re.MULTILINE,
+        )
+        self.assertIsNone(forbidden_import.search(launcher))
+        for function_name in (
+            "generate_mask",
+            "connected_descriptor",
+            "generate_geometry",
+            "parse_args",
+            "main",
+        ):
+            self.assertIn(f"def {function_name}(", runtime)
+        self.assertIn(
+            'EXPECTED_PUCKWORKS_COMMIT = "fc61c4670ec7bf801e40bb391aab16048b8da26b"',
+            runtime,
+        )
+        self.assertIn("Path(__file__).resolve().parents[3]", runtime)
+        self.assertIn("sys.dont_write_bytecode = True", runtime)
+        geometry_manifest = CASE_ROOT / "XSV_TAICHI_001_GEOMETRY_MANIFEST.json"
+        self.assertEqual(
+            hashlib.sha256(geometry_manifest.read_bytes()).hexdigest(),
+            "5ddb9617b3543d7f48eecf5941291d265894a6cd2d5a142265a0750ab509afdd",
         )
 
     def test_exact_prospective_run_matrices(self) -> None:
@@ -98,8 +141,16 @@ class XSVTaichi001Tests(unittest.TestCase):
             "SP32": "40196fd2f2b86de853f2afcfce801b6da0fca1d399e107c6ed40328776ed5a85",
             "M0A": "10d9a010cbac4b8579154456c4271ecd2808af5116beab15a2ffd4e2c99cd039",
         }
+        expected_configs = {
+            "CH33": "674b2f00791e0f12f9dd5cf8c26b98afd56f2ccde81980189cfddda7c5afbebb",
+            "SP32": "6f086749b44555938e2b5612c2a41ebcc6536661756a541a78f5c1657d6eb9ac",
+            "M0A": "4c5ded6230ec5781c810d2a8f53a92a507a3b393cc99f16d9c2a8e5fb027b4f9",
+        }
         for case_id, row in geometries.items():
             self.assertEqual(row["payload_sha256"], expected_payloads[case_id])
+            self.assertEqual(
+                row["geometry_config_sha256"], expected_configs[case_id]
+            )
             self.assertTrue(row["x_through_connected"])
             self.assertGreater(row["phi_gross"], 0.0)
             self.assertLess(row["phi_gross"], 1.0)
